Refactor triangulation into free functions
