+++ tests/Helpers.cpp
@@ -7,7 +7,7 @@
 #include "pmp/algorithms/Remeshing.h"
 #include "pmp/algorithms/Features.h"
 #include "pmp/algorithms/subdivision.h"
-#include "pmp/algorithms/Triangulation.h"
+#include "pmp/algorithms/triangulation.h"
 
 namespace pmp {
