+++ tests/RemeshingTest.cpp
@@ -6,7 +6,7 @@
 #include "pmp/algorithms/Remeshing.h"
 #include "pmp/algorithms/Features.h"
 #include "pmp/algorithms/shapes.h"
-#include "pmp/algorithms/Triangulation.h"
+#include "pmp/algorithms/triangulation.h"
 #include "pmp/utilities.h"
 
 #include "Helpers.h"
@@ -17,7 +17,7 @@ using namespace pmp;
 TEST(RemeshingTest, adaptive_remeshing_with_features)
 {
     auto mesh = cylinder();
-    Triangulation(mesh).triangulate();
+    triangulate(mesh);
     Features(mesh).detect_angle(25);
     auto bb = bounds(mesh).size();
     Remeshing(mesh).adaptive_remeshing(0.001 * bb,  // min length
